Move vault impl to a single file
